# Optimise `BeginAbsoluteSequence` recursive case