Tackling remaining super hot allocations during gameplay (based on number of objects allocated, not size of objects).

master:
|                     Method |       Mean |    Error |   StdDev |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|--------------------------- |-----------:|---------:|---------:|--------:|-------:|------:|----------:|
|               NonRecursive | 1,032.7 us | 19.79 us | 19.44 us |       - |      - |     - |     198 B |
|                  Recursive |   994.2 us | 17.76 us | 16.61 us |  3.9063 |      - |     - |  56,294 B |
| SlightlyNestedNonRecursive | 1,025.9 us | 18.20 us | 16.13 us |       - |      - |     - |     198 B |
|    SlightlyNestedRecursive | 1,185.3 us | 23.11 us | 20.49 us | 17.5781 | 5.8594 |     - | 208,294 B |
|     VeryNestedNonRecursive | 1,019.5 us | 20.12 us | 16.80 us |       - |      - |     - |     194 B |
|        VeryNestedRecursive | 1,108.5 us | 21.74 us | 22.32 us | 11.7188 | 1.9531 |     - | 152,294 B |

this pr:
|                     Method |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |-----------:|---------:|---------:|-------:|------:|------:|----------:|
|               NonRecursive | 1,019.7 us | 19.81 us | 18.53 us |      - |     - |     - |     194 B |
|                  Recursive |   975.4 us | 18.79 us | 25.09 us | 3.9063 |     - |     - |  56,291 B |
| SlightlyNestedNonRecursive | 1,023.6 us | 18.99 us | 17.77 us |      - |     - |     - |     198 B |
|    SlightlyNestedRecursive | 1,090.3 us | 17.23 us | 15.28 us | 9.7656 |     - |     - | 120,334 B |
|     VeryNestedNonRecursive | 1,034.7 us | 14.08 us | 13.17 us |      - |     - |     - |     194 B |
|        VeryNestedRecursive | 1,026.4 us | 19.24 us | 20.59 us | 3.9063 |     - |     - |  64,870 B |

Is this direction is seen as agreeable, the same refactor can be applied to `BeginDelayedSequence`. There is probably more that can (and should) be done with transform logic, but this is a quick win.