@peppy 2021-08-18
Real world:

single play session (running through `TestSceneAutoplay` once) 

object alloc count reduction 15%
memory alloc bytes reduction 10%

![image](https://user-images.githubusercontent.com/191335/129928923-3fe2db3e-8ffb-4124-831c-cef8b9b5f62d.png)